Separate Segment, reorganize directory structure
