+++ src/containers/MainContainer.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import FormContainer from './FormContainer'
 import CanvasContainer from './CanvasContainer'
-import NewCanvasContainer from './NewCanvasContainer'
 import '../styles/style.scss'
 
 export default class MainContainer extends React.PureComponent {
@@ -13,19 +12,9 @@ export default class MainContainer extends React.PureComponent {
         baseSv: 1,
         svMultiplier: 1
       },
-      gridSize: 4,
-      points: []
+      gridSize: 4
     }
   }
-
-      //   <CanvasContainer
-      //   key="canvas_container"
-      //   grid={this.state.grid}
-      //   points={this.state.points}
-      //   onAddPoint={this.handleAddPoint}
-      //   onChangePoint={this.handleChangePoint}
-      //   onDeletePoint={this.handleDeletePoint}
-      // />,
 
   render() {
     return [
@@ -33,13 +22,9 @@ export default class MainContainer extends React.PureComponent {
         key="form_container"
         onChangeSettings={this.handleChangeSettings}
       />,
-      <NewCanvasContainer
+      <CanvasContainer
         key="canvas_container"
         gridSize={this.state.gridSize}
-        points={this.state.points}
-        onAddPoint={this.handleAddPoint}
-        onChangePoint={this.handleChangePoint}
-        onDeletePoint={this.handleDeletePoint}
       />
     ]
   }
@@ -48,42 +33,7 @@ export default class MainContainer extends React.PureComponent {
     this.setState({ settings })
   }
 
-  handleChangeGrid = (grid) => {
-    this.setState({ grid })
-  }
-
-  handleAddPoint = (point, index) => {
-    this.setState((currentState) => {
-      const points = currentState.points
-      const i = index || points.length
-      const lastPt = points[i - 1] || {}
-      const dx = Math.abs(point.x - lastPt.x)
-      const dy = Math.abs(point.y - lastPt.y)
-      if (dx < 4 && dy < 4) {
-        lastPt.anchor = true
-      } else {
-        points.splice(i, 0, point)
-      }
-      return currentState
-    })
-  }
-
-  handleChangePoint = (point, index) => {
-    this.setState((currentState) => {
-      currentState.points[index] = point
-      return currentState
-    })
-  }
-
-  handleDeletePoint = (index) => {
-    this.setState((currentState) => {
-      const point = currentState.points[index]
-      if (point.anchor) {
-        point.anchor = false
-      } else {
-        currentState.points.splice(index, 1)
-      }
-      return currentState
-    })
+  handleChangeGridSize = (gridSize) => {
+    this.setState({ gridSize })
   }
 }
